cleaner code: split code into smaller logical units/classes
Forecast in Material Disposition
https://github.com/metasfresh/metasfresh/issues/2643
