+++ de.metas.material/dispo-service/src/test/java/de/metas/material/dispo/service/event/MDEventListenerTests.java
@@ -8,7 +8,6 @@ import static org.junit.Assert.assertThat;
 
 import java.math.BigDecimal;
 import java.util.Date;
-import java.util.List;
 
 import org.adempiere.test.AdempiereTestHelper;
 import org.adempiere.test.AdempiereTestWatcher;
@@ -28,12 +27,14 @@ import de.metas.material.dispo.DispoTestUtils;
 import de.metas.material.dispo.model.I_MD_Candidate;
 import de.metas.material.dispo.service.CandidateChangeHandler;
 import de.metas.material.dispo.service.StockCandidateFactory;
-import de.metas.material.dispo.service.event.DistributionPlanEventHandler;
-import de.metas.material.dispo.service.event.MDEventListener;
-import de.metas.material.dispo.service.event.ProductionPlanEventHandler;
-import de.metas.material.dispo.service.event.SupplyProposalEvaluator;
+import de.metas.material.dispo.service.event.handler.DistributionPlanEventHandler;
+import de.metas.material.dispo.service.event.handler.ForecastEventHandler;
+import de.metas.material.dispo.service.event.handler.ProductionPlanEventHandler;
+import de.metas.material.dispo.service.event.handler.ReceiptScheduleEventHandler;
+import de.metas.material.dispo.service.event.handler.ShipmentScheduleEventHandler;
+import de.metas.material.dispo.service.event.handler.ShipmentScheduleEventHandlerTests;
+import de.metas.material.dispo.service.event.handler.TransactionEventHandler;
 import de.metas.material.event.EventDescr;
-import de.metas.material.event.MaterialDescriptor;
 import de.metas.material.event.MaterialEventService;
 import de.metas.material.event.ShipmentScheduleEvent;
 import de.metas.material.event.ddorder.DDOrder;
@@ -71,8 +72,6 @@ import mockit.Mocked;
  */
 public class MDEventListenerTests
 {
-	private static final int orderLineId = 86;
-
 	/** Watches the current tests and dumps the database to console in case of failure */
 	@Rule
 	public final TestWatcher testWatcher = new AdempiereTestWatcher();
@@ -105,10 +104,12 @@ public class MDEventListenerTests
 		final CandidateRepository candidateRepository = new CandidateRepository();
 		final SupplyProposalEvaluator supplyProposalEvaluator = new SupplyProposalEvaluator(candidateRepository);
 
-		final CandidateChangeHandler candidateChangeHandler = new CandidateChangeHandler(candidateRepository, new StockCandidateFactory(candidateRepository), materialEventService);
+		final StockCandidateFactory stockCandidateService = new StockCandidateFactory(candidateRepository);
+
+		final CandidateChangeHandler candidateChangeHandler = new CandidateChangeHandler(candidateRepository, stockCandidateService, materialEventService);
 
 		final CandidateService candidateService = new CandidateService(
-				candidateRepository, 
+				candidateRepository,
 				MaterialEventService.createLocalServiceThatIsReadyToUse());
 
 		final DistributionPlanEventHandler distributionPlanEventHandler = new DistributionPlanEventHandler(
@@ -121,46 +122,19 @@ public class MDEventListenerTests
 
 		final ForecastEventHandler forecastEventHandler = new ForecastEventHandler(candidateChangeHandler);
 
+		final TransactionEventHandler transactionEventHandler = new TransactionEventHandler(stockCandidateService, candidateChangeHandler);
+
+		final ShipmentScheduleEventHandler shipmentScheduleEventHandler = new ShipmentScheduleEventHandler(candidateChangeHandler);
+
+		final ReceiptScheduleEventHandler receiptScheduleEventHandler = new ReceiptScheduleEventHandler(candidateChangeHandler);
+
 		mdEventListener = new MDEventListener(
-				candidateChangeHandler,
 				distributionPlanEventHandler,
 				productionPlanEventHandler,
-				forecastEventHandler);
-	}
-
-	/**
-	 * This test is more for myself, to figure out how the system works :-$
-	 */
-	@Test
-	public void testShipmentScheduleEvent()
-	{
-		final ShipmentScheduleEvent event = ShipmentScheduleEvent.builder()
-				.eventDescr(new EventDescr(org.getAD_Client_ID(), org.getAD_Org_ID()))
-				.materialDescr(MaterialDescriptor.builder()
-						.date(t1)
-						.productId(productId)
-						.qty(BigDecimal.TEN)
-						.warehouseId(toWarehouseId)
-						.build())
-				.reference(TableRecordReference.of("someTable", 4))
-				.orderLineId(orderLineId)
-				.build();
-		mdEventListener.onEvent(event);
-
-		final List<I_MD_Candidate> allRecords = DispoTestUtils.retrieveAllRecords();
-		assertThat(allRecords.size(), is(2));
-
-		assertThat(DispoTestUtils.filter(Type.DEMAND).size(), is(1));
-		assertThat(DispoTestUtils.filter(Type.STOCK).size(), is(1));
-
-		final I_MD_Candidate demandRecord = DispoTestUtils.filter(Type.DEMAND).get(0);
-		final I_MD_Candidate stockRecord = DispoTestUtils.filter(Type.STOCK).get(0);
-
-		assertThat(demandRecord.getSeqNo(), is(stockRecord.getSeqNo() - 1)); // the demand record shall be displayed first
-		assertThat(stockRecord.getMD_Candidate_Parent_ID(), is(demandRecord.getMD_Candidate_ID()));
-
-		assertThat(demandRecord.getQty(), comparesEqualTo(BigDecimal.TEN));
-		assertThat(stockRecord.getQty(), comparesEqualTo(BigDecimal.TEN.negate())); // the stock is unbalanced, because there is no existing stock and no supply
+				forecastEventHandler,
+				transactionEventHandler,
+				receiptScheduleEventHandler,
+				shipmentScheduleEventHandler);
 	}
 
 	/**
@@ -169,7 +143,8 @@ public class MDEventListenerTests
 	@Test
 	public void testShipmentScheduleEvent_then_DistributionPlanevent()
 	{
-		testShipmentScheduleEvent();
+		final ShipmentScheduleEvent shipmentScheduleEvent = ShipmentScheduleEventHandlerTests.createShipmentScheduleTestEvent(org);
+		mdEventListener.onEvent(shipmentScheduleEvent);
 
 		// create a DistributionPlanEvent event which matches the shipmentscheduleEvent that we processed in testShipmentScheduleEvent()
 		final TableRecordReference reference = TableRecordReference.of("someTable", 4);
